+++ packages/zarm/src/cell/index.tsx
@@ -1,16 +1,14 @@
 import React, { HTMLAttributes } from 'react';
 import classnames from 'classnames';
+import { ConfigContext } from '../n-config-provider';
 import type { BaseCellProps } from './interface';
 
 type HTMLDivProps = Omit<HTMLAttributes<HTMLDivElement>, 'title'>;
 
-export interface CellProps extends HTMLDivProps, BaseCellProps {
-  prefixCls?: string;
-}
+export type CellProps = HTMLDivProps & BaseCellProps;
 
 const Cell = React.forwardRef<unknown, CellProps>((props, ref) => {
   const {
-    prefixCls,
     className,
     hasArrow,
     icon,
@@ -23,7 +21,10 @@ const Cell = React.forwardRef<unknown, CellProps>((props, ref) => {
     ...others
   } = props;
 
-  const cellRef = (ref as any) || React.createRef<HTMLElement>();
+  const cellRef = (ref as any) || React.createRef<HTMLDivElement>();
+
+  const { prefixCls: globalPrefixCls } = React.useContext(ConfigContext);
+  const prefixCls = `${globalPrefixCls}-cell`;
 
   const cls = classnames(prefixCls, className, {
     [`${prefixCls}--disabled`]: disabled,
@@ -60,7 +61,6 @@ const Cell = React.forwardRef<unknown, CellProps>((props, ref) => {
 Cell.displayName = 'Cell';
 
 Cell.defaultProps = {
-  prefixCls: 'za-cell',
   hasArrow: false,
   disabled: false,
 };