+++ packages/zarm/src/custom-input/CustomInput.tsx
@@ -4,17 +4,14 @@ import { CloseCircleFill } from '@zarm-design/icons';
 import KeyboardPicker from '../keyboard-picker';
 import useClickAway from '../useClickAway';
 import { getValue } from '../input/utils';
+import { ConfigContext } from '../n-config-provider';
 import type { BaseCustomInputProps } from './interface';
 
-export interface CustomInputProps
-  extends BaseCustomInputProps,
-    Omit<React.InputHTMLAttributes<HTMLInputElement>, 'type' | 'onChange' | 'onFocus' | 'onBlur'> {
-  prefixCls?: string;
-}
+export type CustomInputProps = BaseCustomInputProps &
+  Omit<React.InputHTMLAttributes<HTMLInputElement>, 'type' | 'onChange' | 'onFocus' | 'onBlur'>;
 
 const CustomInput = React.forwardRef<unknown, CustomInputProps>((props, ref) => {
   const {
-    prefixCls,
     type,
     clearable,
     readOnly,
@@ -31,9 +28,9 @@ const CustomInput = React.forwardRef<unknown, CustomInputProps>((props, ref) => 
     ...restProps
   } = props;
 
-  const wrapperRef = (ref as any) || React.createRef<HTMLElement>();
+  const wrapperRef = (ref as any) || React.createRef<HTMLDivElement>();
   const contentRef = React.useRef<HTMLDivElement>(null);
-  const pickerRef = React.useRef<HTMLDivElement>();
+  const pickerRef = React.useRef<HTMLDivElement>(null);
   const inputRef = React.useRef<HTMLInputElement>(null);
   const [currentValue, setCurrentValue] = React.useState(getValue({ value, defaultValue }, ''));
   const [focused, setFocused] = React.useState<boolean>(autoFocus!);
@@ -44,6 +41,9 @@ const CustomInput = React.forwardRef<unknown, CustomInputProps>((props, ref) => 
     typeof value !== 'undefined' &&
     currentValue.length > 0 &&
     typeof onChange === 'function';
+
+  const { prefixCls: globalPrefixCls } = React.useContext(ConfigContext);
+  const prefixCls = `${globalPrefixCls}-custom-input`;
 
   const cls = classnames(prefixCls, `${prefixCls}--${type}`, className, {
     [`${prefixCls}--disabled`]: disabled,
@@ -167,7 +167,6 @@ const CustomInput = React.forwardRef<unknown, CustomInputProps>((props, ref) => 
 CustomInput.displayName = 'CustomInput';
 
 CustomInput.defaultProps = {
-  prefixCls: 'za-custom-input',
   type: 'number',
   clearable: true,
   readOnly: false,