+++ packages/zarm/src/custom-input/interface.ts
@@ -3,7 +3,7 @@ export type CustomInputType = 'number' | 'price' | 'idcard';
 export interface BaseCustomInputProps {
   type?: CustomInputType;
   clearable?: boolean;
-  onChange?: (value?: string) => void;
-  onFocus?: (value?: string) => void;
-  onBlur?: (value?: string) => void;
+  onChange?: (value: string) => void;
+  onFocus?: (value: string) => void;
+  onBlur?: (value: string) => void;
 }