+++ packages/zarm/src/locale-provider/locale/zh_CN.tsx
@@ -1,3 +1,0 @@
-import locale from '../../config-provider/locale/zh_CN';
-
-export default locale;