+++ packages/zarm/src/message/Message.tsx
@@ -1,16 +1,14 @@
-import React, { forwardRef, useState } from 'react';
+import React, { forwardRef, useState, useContext } from 'react';
 import type { HTMLAttributes, MouseEvent } from 'react';
 import classnames from 'classnames';
 import { ArrowRight as ArrowRightIcon, Close as CloseIcon } from '@zarm-design/icons';
+import { ConfigContext } from '../n-config-provider';
 import type { BaseMessageProps } from './interface';
 
-export interface MessageProps extends BaseMessageProps, HTMLAttributes<HTMLDivElement> {
-  prefixCls?: string;
-}
+export type MessageProps = BaseMessageProps & HTMLAttributes<HTMLDivElement>;
 
 const Message = forwardRef<HTMLDivElement, MessageProps>((props, ref) => {
   const {
-    prefixCls,
     className,
     theme,
     size,
@@ -22,7 +20,11 @@ const Message = forwardRef<HTMLDivElement, MessageProps>((props, ref) => {
     ...restProps
   } = props;
   const [visible, setVisible] = useState(true);
+  const { prefixCls: globalPrefixCls } = useContext(ConfigContext);
 
+  if (!visible) return null;
+
+  const prefixCls = `${globalPrefixCls}-message`;
   const cls = classnames(prefixCls, className, {
     [`${prefixCls}--${theme}`]: theme,
     [`${prefixCls}--${size}`]: size,
@@ -37,37 +39,27 @@ const Message = forwardRef<HTMLDivElement, MessageProps>((props, ref) => {
     onClick?.(e);
   }
 
-  function handleClose() {
-    setVisible(false);
-  }
-
-  const iconRender = icon && <div className={`${prefixCls}__icon`}>{icon}</div>;
-  const closeIconRender = closable && <CloseIcon onClick={handleClose} />;
-  const arrowRender = hasArrow && <ArrowRightIcon />;
-  const noFooter = !closable && !hasArrow;
-
-  if (visible) {
-    return (
-      <div ref={ref} className={cls} onClick={handleClick} {...restProps}>
-        <div className={`${prefixCls}__header`}>{iconRender}</div>
-        <div className={`${prefixCls}__body`}>{children}</div>
-        {!noFooter && (
-          <div className={`${prefixCls}__footer`}>
-            {arrowRender}
-            {closeIconRender}
-          </div>
-        )}
-      </div>
-    );
-  }
-
-  return null;
+  return (
+    <div ref={ref} className={cls} onClick={handleClick} {...restProps}>
+      {icon && (
+        <div className={`${prefixCls}__header`}>
+          <div className={`${prefixCls}__icon`}>{icon}</div>
+        </div>
+      )}
+      <div className={`${prefixCls}__body`}>{children}</div>
+      {(closable || hasArrow) && (
+        <div className={`${prefixCls}__footer`}>
+          {hasArrow && <ArrowRightIcon />}
+          {closable && <CloseIcon onClick={() => setVisible(false)} />}
+        </div>
+      )}
+    </div>
+  );
 });
 
 Message.displayName = 'Message';
 
 Message.defaultProps = {
-  prefixCls: 'za-message',
   theme: 'primary',
   hasArrow: false,
   closable: false,