+++ packages/zarm/src/tabs/Tabs.tsx
@@ -2,8 +2,10 @@ import * as React from 'react';
 import classnames from 'classnames';
 import TabPanel from './TabPanel';
 import Carousel from '../carousel';
+import type { CarouselHTMLElement } from '../carousel';
 import { getTransformPropValue, getPxStyle } from './util/index';
 import { scrollTo } from '../utils/dom';
+import { ConfigContext } from '../n-config-provider';
 import type { TabPanelProps } from './TabPanel';
 import type { BaseTabsProps } from './interface';
 
@@ -45,18 +47,16 @@ const getValue = (props: TabsProps, defaultValue: TabsProps['value']) => {
 };
 
 export interface TabsProps extends BaseTabsProps {
-  prefixCls?: string;
   className?: string;
 }
 
 interface CompoundedComponent
-  extends React.ForwardRefExoticComponent<TabsProps & React.RefAttributes<HTMLElement>> {
+  extends React.ForwardRefExoticComponent<TabsProps & React.RefAttributes<HTMLDivElement>> {
   Panel: typeof TabPanel;
 }
 
 const Tabs = React.forwardRef<unknown, TabsProps>((props, ref) => {
   const {
-    prefixCls,
     className,
     value,
     defaultValue,
@@ -69,13 +69,16 @@ const Tabs = React.forwardRef<unknown, TabsProps>((props, ref) => {
     children,
   } = props;
 
-  const tabsRef = (ref as any) || React.createRef<HTMLElement>();
-  const carouselRef = React.useRef<Carousel>(null);
+  const tabsRef = (ref as any) || React.createRef<HTMLDivElement>();
+  const carouselRef = React.useRef<CarouselHTMLElement>(null);
   const tablistRef = React.useRef<HTMLUListElement>(null);
   const [itemWidth, setItemWidth] = React.useState(0);
   const [currentValue, setCurrentValue] = React.useState(
     getValue({ value, defaultValue, children }, 0),
   );
+
+  const { prefixCls: globalPrefixCls } = React.useContext(ConfigContext);
+  const prefixCls = `${globalPrefixCls}-tabs`;
 
   const isVertical: boolean = direction === 'vertical';
 
@@ -243,7 +246,6 @@ const Tabs = React.forwardRef<unknown, TabsProps>((props, ref) => {
 Tabs.displayName = 'Tabs';
 
 Tabs.defaultProps = {
-  prefixCls: 'za-tabs',
   disabled: false,
   swipeable: false,
   scrollable: false,